pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "duckduckgo-search>=7.5.5",
    "html2text>=2024.2.26",
    "langchain>=0.3.22",
    "langchain-core>=0.3.49",
    "langchain-groq>=0.3.2",
    "langgraph>=0.3.21",
    "pandas>=2.2.3",
    "plotly>=6.0.1",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "streamlit>=1.44.0",
    "trafilatura>=2.0.0",
]
